Refactor Of Header Files
Split up headers, predefine where we can, and use precompiled headers to speed up compile times. Includes some CString/QString changes to aid in the above where only a simple change was needed.

+++ Tools/ButtonToolbar.cpp
@@ -13,9 +13,9 @@ Bitmap *	GetBitmapFromIcon(int nBitmap, int nMask)
 	Bitmap				bmpMask(hMask, nullptr);
 
 	pResult = bmp.Clone(0, 0, BUTTONTOOLBARSIZE, BUTTONTOOLBARSIZE, PixelFormat32bppARGB);
-	for (LONG i = 0;i<bmp.GetWidth();i++)
+	for (UINT i = 0;i<bmp.GetWidth();i++)
 	{
-		for (LONG j = 0;j<bmp.GetHeight();j++)
+		for (UINT j = 0;j<bmp.GetHeight();j++)
 		{
 			Color		color,
 						tmpcolor;
